Use ENV instead of flags to pass in args to tests
